requirements.txt
streamlit
google-generativeai
pymupdf
python-dotenv
flask
flask-cors

